+++ src/applications/income-and-asset-statement/config/chapters/05-royalties-and-other-properties/index.js
